backend/agents/requirements.txt
python-docx
openai
pandas
python-dotenv
pytesseract
pdf2image
google-generativeai
opencv-python
Pillow
pydub
torchaudio
transformers
soundfile 
noisereduce
streamlit
